move some duplication around
