+++ plugin-server/src/main/ingestion-queues/session-recording/services/console-logs-ingester.ts
@@ -1,15 +1,14 @@
 import { captureException } from '@sentry/node'
-import { HighLevelProducer as RdKafkaProducer, NumberNullUndefined } from 'node-rdkafka'
+import { HighLevelProducer as RdKafkaProducer } from 'node-rdkafka'
 import { Counter } from 'prom-client'
 
 import { KAFKA_LOG_ENTRIES } from '../../../../config/kafka-topics'
-import { findOffsetsToCommit } from '../../../../kafka/consumer'
-import { retryOnDependencyUnavailableError } from '../../../../kafka/error-handling'
-import { flushProducer, produce } from '../../../../kafka/producer'
+import { produce } from '../../../../kafka/producer'
 import { status } from '../../../../utils/status'
 import { eventDroppedCounter } from '../../metrics'
 import { ConsoleLogEntry, gatherConsoleLogEvents, RRWebEventType } from '../process-event'
 import { IncomingRecordingMessage } from '../types'
+import { BaseIngester } from './base-ingester'
 import { OffsetHighWaterMarker } from './offset-high-water-marker'
 
 const HIGH_WATERMARK_KEY = 'session_replay_console_logs_events_ingester'
@@ -37,70 +36,9 @@ function deduplicateConsoleLogEvents(consoleLogEntries: ConsoleLogEntry[]): Cons
     return deduped
 }
 
-// TODO this is an almost exact duplicate of the replay events ingester
-// am going to leave this duplication and then collapse it when/if we add a performance events ingester
-export class ConsoleLogsIngester {
-    constructor(
-        private readonly producer: RdKafkaProducer,
-        private readonly persistentHighWaterMarker?: OffsetHighWaterMarker
-    ) {}
-
-    public async consumeBatch(messages: IncomingRecordingMessage[]) {
-        const pendingProduceRequests: Promise<NumberNullUndefined>[] = []
-
-        for (const message of messages) {
-            const results = await retryOnDependencyUnavailableError(() => this.consume(message))
-            if (results) {
-                pendingProduceRequests.push(...results)
-            }
-        }
-
-        // On each loop, we flush the producer to ensure that all messages
-        // are sent to Kafka.
-        try {
-            await flushProducer(this.producer!)
-        } catch (error) {
-            // Rather than handling errors from flush, we instead handle
-            // errors per produce request, which gives us a little more
-            // flexibility in terms of deciding if it is a terminal
-            // error or not.
-        }
-
-        // We wait on all the produce requests to complete. After the
-        // flush they should all have been resolved/rejected already. If
-        // we get an intermittent error, such as a Kafka broker being
-        // unavailable, we will throw. We are relying on the Producer
-        // already having handled retries internally.
-        for (const produceRequest of pendingProduceRequests) {
-            try {
-                await produceRequest
-            } catch (error) {
-                status.error('🔁', '[console-log-events-ingester] main_loop_error', { error })
-
-                if (error?.isRetriable) {
-                    // We assume that if the error is retriable, then we
-                    // are probably in a state where e.g. Kafka is down
-                    // temporarily, and we would rather simply throw and
-                    // have the process restarted.
-                    throw error
-                }
-            }
-        }
-
-        if (this.persistentHighWaterMarker) {
-            const topicPartitionOffsets = findOffsetsToCommit(
-                messages.map((message) => ({
-                    topic: message.metadata.topic,
-                    partition: message.metadata.partition,
-                    offset: message.metadata.highOffset,
-                }))
-            )
-            await Promise.all(
-                topicPartitionOffsets.map((tpo) =>
-                    this.persistentHighWaterMarker!.add(tpo, HIGH_WATERMARK_KEY, tpo.offset)
-                )
-            )
-        }
+export class ConsoleLogsIngester extends BaseIngester {
+    constructor(producer: RdKafkaProducer, persistentHighWaterMarker?: OffsetHighWaterMarker) {
+        super('session_replay_console_logs_events_ingester', producer, persistentHighWaterMarker)
     }
 
     public async consume(event: IncomingRecordingMessage): Promise<Promise<number | null | undefined>[] | void> {
